lapa/frontends/cpp.py
"""
C/C++ language frontend for LAPA framework.

This module provides the implementation of a C/C++ language frontend that
parses C and C++ code into the framework's IR using LLVM/Clang.
"""

import os
import sys
from pathlib import Path
from typing import Any, Dict, List, Optional, Union, cast

from ..frontend import LanguageFeatures, LanguageFrontend, ParsingError
from ..ir import IR, IRNode, IRNodeType, Position


class CPPFrontend(LanguageFrontend):
    """Frontend implementation for C/C++."""

    def __init__(self):
        """Initialize the C/C++ frontend."""
        super().__init__()
        # TODO: Initialize LLVM/Clang integration
        self.parser = None

    def _get_language_features(self) -> LanguageFeatures:
        """Get C/C++ language features."""
        features = LanguageFeatures()
        features.has_classes = True  # C++ only
        features.has_interfaces = False
        features.has_generics = True  # C++ templates
        features.has_exceptions = True  # C++ only
        features.has_async = False
        features.has_decorators = False
        features.has_operator_overloading = True  # C++ only
        features.has_multiple_inheritance = True  # C++ only
        features.typing_system = "static"
        features.memory_management = "manual"
        return features

    def parse_file(self, path: Union[str, Path]) -> IR:
        """Parse a C/C++ file into IR."""
        path = Path(path)
        if not path.exists():
            raise FileNotFoundError(f"File not found: {path}")
        
        try:
            with open(path, 'r', encoding='utf-8') as f:
                content = f.read()
            return self.parse_string(content, str(path))
        except Exception as e:
            raise ParsingError(f"Failed to parse {path}: {str(e)}")

    def parse_string(self, content: str, filename: str = "<string>") -> IR:
        """Parse C/C++ code from a string into IR."""
        raise NotImplementedError(
            "C/C++ parsing not yet implemented"
        )

    def get_file_extensions(self) -> List[str]:
        """Get supported C/C++ file extensions."""
        return [".c", ".h", ".cpp", ".hpp", ".cc", ".hh", ".cxx", ".hxx"]

    def _process_tree(self, tree: Any, filename: str) -> None:
        """Process AST and convert it to IR."""
        # TODO: Implement AST processing
        pass

    def _process_node(self, node: Any, parent_ir: IRNode) -> None:
        """Process an AST node and its children."""
        # TODO: Implement node processing
        pass

    def _process_function(self, node: Any, parent_ir: IRNode) -> None:
        """Process a function declaration."""
        # TODO: Implement function processing
        pass

    def _process_class(self, node: Any, parent_ir: IRNode) -> None:
        """Process a class declaration."""
        # TODO: Implement class processing
        pass

    def _process_variable(self, node: Any, parent_ir: IRNode) -> None:
        """Process a variable declaration."""
        # TODO: Implement variable processing
        pass

    def _process_include(self, node: Any, parent_ir: IRNode) -> None:
        """Process an include directive."""
        # TODO: Implement include processing
        pass

    def _process_template(self, node: Any, parent_ir: IRNode) -> None:
        """Process a template declaration."""
        # TODO: Implement template processing
        pass

    def _create_position(self, node: Any) -> Position:
        """Create Position object from AST node."""
        # TODO: Implement position creation
        return Position(
            line=1,
            column=0,
            file=self.ir.root.position.file if self.ir.root.position else "<unknown>"
        )
